refactor(sgwallet): change local_state_storage to channel_manager
