+++ sgwallet/channel-manager/src/channel.rs
