+++ sgwallet/channel-manager/src/channel_state_view.rs
