+++ sgwallet/channel-manager/src/tests/mod.rs
@@ -1,5 +1,5 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-mod local_state_storage_test;
+mod channel_manager_test;
 mod tx_applier_test;
